requirements.txt
# --- Frontend & UI ---
streamlit>=1.31.0

# --- AI Orchestration & RAG ---
langchain>=0.1.0
langchain-community>=0.0.20
langchain-ollama>=0.1.0
langchain-huggingface>=0.0.1
langchain-chroma>=0.1.0

# --- Vector Database & Embeddings ---
chromadb>=0.4.24
sentence-transformers>=2.5.0
huggingface-hub>=0.20.0

# --- Document Processing ---
pypdf>=4.0.0
python-docx>=1.1.0
fpdf2>=2.7.0

# --- NLP & Analytics ---
textblob>=0.17.1

# --- Utilities ---
python-dotenv>=1.0.1
tiktoken>=0.5.0